+++ CHANGES.md
@@ -5,6 +5,7 @@ Changelog
 ---------
 
 - Fixed a bug checking levels across inactive items.
+- Added error message for all IO errors.
 
 0.8.3 (2014/10/10)
 ------------------